New Critical Updates

+++ src/Events/messageDelete.js
@@ -4,16 +4,17 @@ module.exports = class Event {
     }
     async run(message) {
         if (message.author.bot) return  // return if msg author is bot
-        if(!message.content) return // return if empty message
-          
-        message.channel.createWebhook((message.member.nickname || message.member.user.username)
-            + ` [${message.translate("common:messageDeleted")}]`, {
-            avatar: message.author.avatarURL({ format: "png"}),
-        })
-            .then(async webhook => {
-                await webhook.send({ content: message.content})
-                await webhook.delete()
-            })
-            .catch((error) => console.error(error));   
+        if (!message.content) return // return if empty message
+
+        const SDM = await this.Xeow.Configuration.get("feature")?.ShowDeletedMessage
+        if (SDM && SDM === true) {
+            message.channel.createWebhook((message.member.nickname || message.member.user.username)
+                + ` [${message.translate("common:messageDeleted")}]`, {
+                avatar: message.author.avatarURL({ format: "png" }),
+            }).then(async webhook => {
+                    await webhook.send({ content: message.content })
+                    await webhook.delete()
+                }).catch((error) => console.error(error));
+        }
     }
 }
